+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/data/model/FilterType.kt
@@ -5,4 +5,5 @@ enum class FilterType {
   Select,
   DateRange,
   AutoComplete,
+  Text,
 }